SONARGRADL-107 Add IT test case with Android Gradle Plugin 8.1.1
Add a new configuration to the integration tests  with AGP 8.1.1 (and Gradle 8.2.1).
This new test does not not modify the plugin but it does require the tweaking of some of our tests and test projects to remain compatible with AGP 8 and earlier versions.

The changes to our test projects are described below.

* Add namespace entries to build.gradle files
For all test projects with Android build properties, a namespace has
been explicitly set to work through the breaking change introduced by
AGP 8+[0].
The namespace picked should match the package of the main class declared
in the test project.

* Remove namespaces from AndroidManifest.xml files
AGP 8+ requires namespaces/packages to be defined in build.gradle files
and be removed from AndroidManifest.xml files.

* Disable`nonTransitiveRClass` to keep access to resources in other
  modules
To keep a consistent behavior with AGP 8+[1], we explicitly disable
nonTransitiveRClass in the properties.

* Change build artifacts expectations in integration tests
The list of folders containing test binaries produced with AGP 8+ is
shorter than the one produced by AGP 7.

[0] https://developer.android.com/build/releases/past-releases/agp-8-0-0-release-notes#namespace-dsl
[1] https://developer.android.com/build/releases/past-releases/agp-8-0-0-release-notes#default-changes