SCANGRADLE-247 Resolve class paths in dedicated task
Because (sub-)project class paths should only be resolved with an
exclusive lock on the project, whether at configuration or execution
time, we introduce a new task SonarResolverTask dedicated to this job.

In practice, we register for each project a resolver task that resolves
the compile and test compile class paths and stores them in a designated
(not-shared) output file.
The sonar task can then depend on these files as input and benefit from
the (cacheable) work done ahead by the resolver tasks.

As a consequence of this change: class path dependent sonar properties
such as  sonar.java.libraries and sonar.java.test.libraries are no
longer resolved at configuration time (see changes in related tests).
Execution of the sonar task is now necessary, for these properties to be
effectively computed and finalized.