SCANGRADLE-293 mark SonarResolverTask as UntrackedTask (#392)
If tasks affect classpath gralde will detect that and throw an error relative to the SonarResolverTask, because SonarResolverTask depends on the classpath.

Two solution exits do not use any gradle caching mechanism with UntrackedTask or specify an explicit
relation (depends on, must run after, must run before, ... ) between the tasks.